requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
gunicorn==20.0.4
Flask==2.0.2
google-cloud-storage==1.43.0
Flask-Migrate
Flask-SQLAlchemy
filedepot
marhsmallow-jsonapi
webargs